Rename `TokenName` to `AssetName`.
To create this commit, we perform the following substitution across all
Haskell source files:

```
s/TokenName/AssetName/
```

Of course, this substitution will also affect *any* token that happens
to include `TokenName` as a substring.

In most cases this is *not* a problem, as in general we *do* want to
rename tokens like `mockTokenName` to `mockAssetName`.

However, we *must* take care to preserve our external API and database
schemas.

In particular, in order to avoid inadvertantly changing any of our
database schemas, we also perform the following additional
substitutions:

```
s/txOutAssetName/txOutTokenName/
s/txCollateralOutAssetName/txCollateralOutTokenName/
s/utxoAssetName/utxoAssetName/
```

The above substitutions have the effect of cancelling out a small subset
of the initial substitution, so that database schemas are left unaffected.

To verify this, we can inspect this commit for changes to database
schemas:

```
$ git diff HEAD~ | grep "sql="
  txOutTokenTxId                TxId             sql=tx_id
  txOutTokenTxIndex             Word32           sql=tx_index
  txOutTokenPolicyId            W.TokenPolicyId  sql=token_policy_id
- txOutTokenName                W.TokenName      sql=token_name
+ txOutTokenName                W.AssetName      sql=token_name
  txOutTokenQuantity            W.TokenQuantity  sql=token_quantity
  txCollateralOutTokenTxId      TxId             sql=tx_id
  txCollateralOutTokenPolicyId  W.TokenPolicyId  sql=token_policy_id
- txCollateralOutTokenName      W.TokenName      sql=token_name
+ txCollateralOutTokenName      W.AssetName      sql=token_name
  txCollateralOutTokenQuantity  W.TokenQuantity  sql=token_quantity
  utxoTokenTxId                 TxId             sql=tx_id
  utxoTokenTxIndex              Word32           sql=tx_index
  utxoTokenPolicyId             W.TokenPolicyId  sql=token_policy_id
- utxoTokenName                 W.TokenName      sql=token_name
+ utxoTokenName                 W.AssetName      sql=token_name
  utxoTokenQuantity             W.TokenQuantity  sql=token_quantity
```