Regenerate golden JSON files for type `TokenMap`.
This commit fixes a warning from `hspec-golden-aeson` that the contents
of `{Flat,Nested}TokenMap.json` do not match the JSON that would be
generated if these files were generated from scratch from the stored
seeds.

This warning is NOT caused by any change in the serialisation of
`TokenMap`. In particular, all field names remain unchanged.