Rename function to `guardUTxOConsistency`.
The `With` suffix was originally there because this function used to be
parameterised. However, this function is no longer parameterised, so we
can remove the `With` suffix.

Additionally, since both UTxO sets are now provided by the scope of the
enclosing function, we can shorten the name of the function so that it
is no longer biased toward the "wallet" UTxO set.