Manually tweak e2e game goldens to be valid in Conway
When reproducing in a unit test I got the following two failures:

- DecoderErrorDeserialiseFailure "Shelley Tx" (DeserialiseFailure 117 "TxBody: 'Required Signer Hashes' must be non-empty when supplied")
- DecoderErrorDeserialiseFailure "Shelley Tx" (DeserialiseFailure 42 "TxBody: 'Collateral Inputs' must be non-empty when supplied")

These correspond to the tags 13 and 14 in the Conway binary spec:
https://github.com/IntersectMBO/cardano-ledger/blob/8d7d261dfb6282ab86cad32ec3f1be71db9a3080/eras/conway/impl/cddl-files/conway.cddl#L431-L432

Using cbor.me we can confirm the presence of in the binaries from the
errors:

```
      0D                                # unsigned(13)
      80                                # array(0)
```

So we manaully delete
- `0d80`
- `0e80`
with this meaning (some matches of 0e80 were part of bytestrings, so we
don't delete them)

And then adjust the preceding TxBody length:

```
84                                      # array(4)
   A5                                   # map(5)
```

E.g. `s/A5/A3`.